Moved stat & getdents hypercalls in separate files
